pyproject.toml
[project]
name = "cinereserve"
version = "0.1.0"
description = ""
authors = [
    {name = "LeaoCarvalho",email = "matheus.eu.pessoal@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "django (<6)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "redis (>=7.3.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
